requirements.txt
geopy
pyusb
pyqt5
pyudev
glib